requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
gunicorn==23.0.0
whitenoise==6.8.2
paho-mqtt==2.1.0

